refactor: cross-module architecture pass (v3.1.31)
Third patch of the post-v3.1.28 desloppify rescore series. Closes
both cross_module_architecture findings. No behaviour change.

Genuine smells fixed:
- New src/utils/xp/configCache.ts owns getXPConfig,
  invalidateXPConfigCache, clearXPConfigCache plus the cache map
  and 5-minute TTL. Lives in utils so xpMessageHandler and
  xpVoiceHandler do not have to cross the events ->
  commands/handlers boundary to read XP config
- xpMessageHandler.ts + xpVoiceHandler.ts: getXPConfig import
  flipped to the new util
- commands/handlers/xp/setup.ts: dropped the local cache map +
  getter/invalidator pair; net -28 lines
- commands/handlers/xp/index.ts: dropped the temporary
  clearXPConfigCache / getXPConfig / invalidateXPConfigCache
  re-export block (those were existence proofs of the smell)
- leaderboard.ts + rank.ts: getXPConfig import flipped to the
  new util

Entity inversion:
- New src/typeorm/entities/ticket/routingTypes.ts owns RoutingRule
  and RoutingStrategy. The entity (data shape) no longer imports
  from a runtime utility
- TicketConfig.ts: RoutingStrategy import flipped to ./routingTypes;
  the inline routingRules Array shape now uses RoutingRule[]
- utils/ticket/smartRouter.ts: types imported from the entity side
  and re-exported for back-compat (existing util-side importers
  still resolve, no churn pass needed). Stale 'requires the
  following columns on TicketConfig' docblock replaced with a
  pointer to the new types module

Documented exception:
- CLAUDE.md: new Layering rule section explains the dispatch
  exception. autocomplete.ts, ticket/interactionRoutes.ts,
  application/interactionRoutes.ts, typeFieldsInteraction.ts,
  applicationFieldsInteraction.ts genuinely DO import handler
  functions from commands/handlers/*. That is the whole point of
  those files; do not 'fix' the dependency direction by moving
  handlers to utils. The per-command logic lives with its slash
  command builder, and the dispatcher is the bridge
