+++ src/commands/handlers/xp/leaderboard.ts
@@ -4,7 +4,7 @@ import xpLang from '../../../lang/en/xp.json';
 import { XPUser } from '../../../typeorm/entities/xp/XPUser';
 import { handleInteractionError } from '../../../utils';
 import { lazyRepo } from '../../../utils/database/lazyRepo';
-import { getXPConfig } from './setup';
+import { getXPConfig } from '../../../utils/xp/configCache';
 
 const LEADERBOARD_COLOR = '#FFD700' as ColorResolvable;
 const PAGE_SIZE = 10;
