+++ src/events/xpVoiceHandler.ts
@@ -6,11 +6,11 @@
  */
 
 import type { VoiceState } from 'discord.js';
-import { getXPConfig } from '../commands/handlers/xp/setup';
 import { XPUser } from '../typeorm/entities/xp/XPUser';
 import type { ExtendedClient } from '../types/ExtendedClient';
 import { enhancedLogger, LogCategory } from '../utils';
 import { lazyRepo } from '../utils/database/lazyRepo';
+import { getXPConfig } from '../utils/xp/configCache';
 import { calculateLevel } from '../utils/xp/xpCalculator';
 
 const userRepo = lazyRepo(XPUser);
